# public/swagger-public.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Next Schedule API - Integração N8N",
    "description": "API pública para integrações externas (ex: N8N, Zapier) utilizando API Key.",
    "version": "1.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
  "security": [],
  "paths": {
    "/api/integrations/appointments": {
      "post": {
        "summary": "Create a new appointment",
        "tags": [
          "Appointments"
        ],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "clientId",
                  "professionalId",
                  "date",
                  "time"
                ],
                "properties": {
                  "clientId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "professionalId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-12-25"
                  },
                  "time": {
                    "type": "string",
                    "example": "14:30"
                  },
                  "appointmentPriceInCents": {
                    "type": "integer",
                    "example": 15000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment created successfully"
          },
          "400": {
            "description": "Invalid payload"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Client or Professional not found"
          },
          "409": {
            "description": "Time slot unavailable"
          }
        }
      },
      "get": {
        "summary": "Get appointments by client ID",
        "tags": [
          "Appointments"
        ],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "clientId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of appointments"
          },
          "400": {
            "description": "Missing client ID"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Client not found"
          }
        }
      },
      "delete": {
        "summary": "Cancel an appointment",
        "tags": [
          "Appointments"
        ],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment cancelled successfully"
          },
          "400": {
            "description": "Missing appointment ID"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Appointment not found"
          }
        }
      },
      "put": {
        "summary": "Update an appointment",
        "tags": [
          "Appointments"
        ],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-12-25"
                  },
                  "time": {
                    "type": "string",
                    "example": "14:30"
                  },
                  "professionalId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment updated successfully"
          },
          "400": {
            "description": "Invalid payload or missing ID"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Appointment, Professional or Client not found"
          },
          "409": {
            "description": "Time slot unavailable"
          }
        }
      }
    },
...
  },
...
}